# Avoid calling `fstat` on things we already know are valid sockets.