This avoids a `stat` system call on the critical path which tells us what we already know.