@ioquatix 2021-07-12
In my benchmark, 100,000 new connections, this was accounting for 0.22s or 15% of the total system call overhead:

Before:

```
% time     seconds  usecs/call     calls    errors syscall                                                                                                                                                                                       
------ ----------- ----------- --------- --------- ------------------                                                                                                                                                                            
 27.28    0.426535           4    100000           write                                                                                                                                                                                         
 23.22    0.363011           3    100240           close                                                                                                                                                                                         
 19.46    0.304226           3    100002         2 accept4                                                                                                                                                                                       
 15.47    0.241845           2    100343        30 read                                                                                                                                                                                          
 14.12    0.220783           2    100392         9 newfstatat                                                                                                                                                                                    
  0.37    0.005845           2      2079           mprotect                              
```

After:

```
% time     seconds  usecs/call     calls    errors syscall
------ ----------- ----------- --------- --------- ------------------
 31.18    0.426379           4    100000           write
 26.71    0.365290           3    100240           close
 22.86    0.312546           3    100002         2 accept4
 18.31    0.250348           2    100343        30 read
  0.54    0.007354           3      2079           mprotect
  0.11    0.001463           2       494       260 openat
  0.08    0.001115           2       551       551 readlink
  0.05    0.000735           1       391         9 newfstatat
```