@ioquatix 2021-07-12
It looks like we should be able to defer all of `close`:

```
 26.71    0.365290           3    100240           close
```

We can chuck this on the `io_uring` without any fiber and ignore the result. Would be another performance win.